% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_skjalftalisa_data.R
\name{build_skjalftalisa_query}
\alias{build_skjalftalisa_query}
\title{Helper function for making a JSON POST query from input variables}
\usage{
build_skjalftalisa_query(
  start_time,
  end_time,
  depth_min = 0,
  depth_max = 25,
  size_min = 0,
  size_max = 7,
  magnitude_preference = c("Mlw", "Autmag"),
  event_type = "qu",
  originating_system = "SIL picks",
  area = c(68, 61, -32, -4),
  fields = c("event_id", "lat", "long", "time", "magnitude", "event_type",
    "originating_system")
)
}
\arguments{
\item{fields}{}
}
\description{
Helper function for making a JSON POST query from input variables
}
