% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_jsk_data.R
\name{download_jsk_data}
\alias{download_jsk_data}
\title{Download Icelandic earthquake data}
\usage{
download_jsk_data(year, week)
}
\arguments{
\item{year}{The year for which to fetch data}

\item{week}{The week for which to fetch data}
}
\value{
An sf table object containing all earthquake measurements from http://hraun.vedur.is/ja/ for the selected interval
}
\description{
Download Icelandic earthquake data
}
\examples{
cur_date <- Sys.Date()
cur_year <- lubridate::year(cur_date)
cur_week <- lubridate::week(cur_date)
d <- download_jsk_data(year = cur_year, week = cur_year)
}
