% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_skjalftalisa_data.R
\name{download_skjalftalisa_data}
\alias{download_skjalftalisa_data}
\title{Download newest Icelandic earthquake data from http://skjalftalisa.vedur.is}
\usage{
download_skjalftalisa_data(
  start_time,
  end_time,
  depth_min = 0,
  depth_max = 25,
  size_min = 0,
  size_max = 7,
  magnitude_preference = c("Mlw", "Autmag"),
  event_type = "qu",
  originating_system = "SIL picks",
  area = c(68, 61, -32, -4),
  fields = c("event_id", "lat", "long", "time", "magnitude", "event_type",
    "originating_system")
)
}
\arguments{
\item{fields}{}
}
\description{
Download newest Icelandic earthquake data from http://skjalftalisa.vedur.is
}
\examples{
start_time <- max(jardskjalftar$time)
end_time <- Sys.time()

d <- download_skjalftalisa_data(start_time, end_time)
}
