% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_skjalftalisa_data.R
\name{example_skjalftalisa_query}
\alias{example_skjalftalisa_query}
\title{An example of a JSON body form a POST request to http://skjalftalisa.vedur.is. Useful for debugging.}
\usage{
example_skjalftalisa_query()
}
\value{
A character vector showing the example JSON body
}
\description{
An example of a JSON body form a POST request to http://skjalftalisa.vedur.is. Useful for debugging.
}
