% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_skjalftalisa_data.R
\name{download_skjalftalisa_data}
\alias{download_skjalftalisa_data}
\title{Download newest Icelandic earthquake data from http://skjalftalisa.vedur.is}
\usage{
download_skjalftalisa_data(
  start_time,
  end_time,
  depth_min = 0,
  depth_max = 25,
  size_min = 0,
  size_max = 7,
  magnitude_preference = c("Mlw", "Autmag"),
  event_type = list("qu"),
  originating_system = list("SIL picks"),
  area = c(68, 61, -32, -4),
  fields = c("event_id", "lat", "long", "time", "magnitude", "depth", "event_type",
    "originating_system"),
  max_weeks_per_request = 32,
  request_rate_per_second = 2
)
}
\arguments{
\item{start_time}{A datetime object or a string}

\item{end_time}{A datetime object or a string}

\item{depth_min}{The minimum earthquake depth to fetch (default = 0)}

\item{depth_max}{The maximum earthquake depth to fetch (default = 25)}

\item{size_min}{The minimum earthquake size to fetch (default = 0)}

\item{size_max}{The maximum earthquake size to fetch (default = 7)}

\item{magnitude_preference}{Preferred magnitude measure -"Mlw" or "Autmag"- (Default is "Mlw")}

\item{event_type}{The type of event to fetch (default is "qu")}

\item{originating_system}{The originating system to fetch from (default is "SIL picks")}

\item{area}{Geographical area to fetch from (default is a bounding box around Iceland)}

\item{fields}{The fields to fetch data on from the service.}

\item{max_weeks_per_request}{For large requests (large duration), how many weeks should there be in each smaller sub-request? (Default 16 weeks per request)}

\item{request_rate_per_second}{How many requests to send per second for larger durations? (Default is 2 per second)}
}
\value{
An {sf} table containing magnitude, location and timestamp of recorded earthquakes
}
\description{
Download newest Icelandic earthquake data from http://skjalftalisa.vedur.is
}
\examples{
start_time <- max(jardskjalftar$time)
end_time <- Sys.time()

d <- download_skjalftalisa_data(start_time, end_time)
}
